+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to
 
 ## [Unreleased]
 
+## [1.18.0] - 2025-06-30
+
 ### Added
 
 - 🐛(front) fix button add mail domain
@@ -409,7 +411,8 @@ and this project adheres to
 - ✨(domains) create and manage domains on admin + API
 - ✨(domains) mailbox creation + link to email provisioning API
 
-[unreleased]: https://github.com/suitenumerique/people/compare/v1.17.0...main
+[unreleased]: https://github.com/suitenumerique/people/compare/v1.18.0...main
+[1.18.0]: https://github.com/suitenumerique/people/releases/v1.18.0
 [1.17.0]: https://github.com/suitenumerique/people/releases/v1.17.0
 [1.16.0]: https://github.com/suitenumerique/people/releases/v1.16.0
 [1.15.0]: https://github.com/suitenumerique/people/releases/v1.15.0